updating testing

+++ src/controllers/productController.js
@@ -41,7 +41,9 @@ function getNavBar() {
 }
 
 function getProductCards(products) {
-    
+    if (!products){
+        throw new Error ('El producto esta vacio o nulo')
+    }
     let option;
     let html = `<div class="cardContainer">`;
     if (token){option = 'dashboard'}else{option = 'products'}
@@ -49,9 +51,10 @@ function getProductCards(products) {
         html += `
         <div class="productCard">
         <h2>${product.name}</h2>
-        <img src="${product.image}" alt="${product.name}">        
+        <img src="${product.image}" alt="${product.name}">      
+        <div class="containerBoton">   
         <a href="/${option}/${product._id}" class="boton">Ver detalle</a>    
-           
+        </div>
         </div>           
         `
     }
@@ -59,20 +62,24 @@ function getProductCards(products) {
 }
 
 function getProductOneCard(product) {
-    
+    if(!product){
+        throw new Error ('El producto esta vacio o nulo')
+    }
     let html = `<div class="cardContainer">
         <div class="productCard">
             <h2>${product.name}</h2>
-            <img src="${product.image}" alt="${product.name}">        
+            <img src="${product.image}" alt="${product.name}">
             <p>${product.description}</p>
             <p>Precio: ${product.price}€</p>
             <p>Talla: ${product.size.toString().toUpperCase()}</p>
             <p>Categoria: ${product.category}</p>
             `
     if (token){
-        html += `        
+        html += ` 
+            <div class="containerBoton">       
             <a href="/dashboard/${product._id}/edit" class="boton">actualizar</a>
-            <a href="/dashboard/${product._id}/delete" class="boton">eliminar</a>                          
+            <a href="/dashboard/${product._id}/delete" class="boton">eliminar</a> 
+            </div>                         
          `
     }
     
@@ -97,8 +104,11 @@ const ProductController = {
         try {
             // controlar con dashboard
             const idProduct = req.params.productId;
-            const product = await Product.findById(idProduct);
             
+            const product = await Product.findById(idProduct);
+            if(!product){
+                throw new Error('El producto no exite')
+            }
             const productCards = getProductOneCard(product, token)
 
             const html = htmlHead + getNavBar() + productCards + htmlEnd
@@ -278,7 +288,7 @@ const ProductController = {
             const idProduct = req.params.productId;
             const deletedProduct = await Product.findByIdAndDelete(idProduct)
             if (!deletedProduct) {
-                return res.status(404).json({ mensaje: 'Product with that idProduct not found' })
+                throw new Error('Producto no encontrado')                
             }
             let message = `<h2>Producto eliminado correctamente</h2>`
             html = htmlHead + getNavBar() + message + htmlEnd
@@ -301,4 +311,8 @@ const ProductController = {
 }
 
 
-module.exports = ProductController
+module.exports = {
+    ProductController,
+    getProductOneCard,
+    getProductCards
+}
